Complete email based authentication example
Error handling is limited, but it’s fully working, so I have added documentation for it.

I have removed hooks and references to Passport for now; I would like to come back later and add in support for oAuth services (Facebook, Twitter and Google in particular) and an example of a user editable profile.