Now silently revalidates sessions periodically
This resolves issue #6.

The client how checks 'clientMaxAge', which is set on the server (along side all the other session options) and uses that value to determine if it it should fetch the session data from the server or from the cache on disk.

Note that the cache on disk is shared between tabs, so logging you out on one will cause you also to be logged out on the other pages when you next interact with the site regardless.

This option just ensures the client will never stray too far from the data on the server (e.g. in the event a users session information is updated or forceably expired on the server).