Refactored how user objects are stored for oAuth
This update includes some changes to how user information is stored in sessions on the server.

User profiles and sessions are now decoupled so that sessions now only store a User ID, not a full user object. This refactoring has been done in perperation for adding oAuth support.

There are also some minor changes which improve performance and reliablity around login and logout in edge cases.