Updating now.json config due to problem deploying on now.sh
now.sh is complaining it can't find the 'pages' directory when deploying, if I specify a 'files' directive, which seems to imply it's not just adding to the list of files that needs to be deploy but is an absolute list if specified.

Removing 'files' resolves the issue, but requires that deploys still be deployed with 'now -E'.